chore: prettify code
