+++ src/main/java/com/twilio/oai/TwilioPhpGenerator.java
@@ -2,10 +2,12 @@ package com.twilio.oai;
 
 import com.twilio.oai.api.ApiResources;
 import com.twilio.oai.api.PhpApiResourceBuilder;
+import com.twilio.oai.api.PhpDomainBuilder;
 import com.twilio.oai.common.EnumConstants;
 import com.twilio.oai.common.Utility;
 import com.twilio.oai.resolver.php.*;
 import com.twilio.oai.resource.ResourceMap;
+import com.twilio.oai.template.IApiActionTemplate;
 import com.twilio.oai.template.PhpApiActionTemplate;
 import io.swagger.v3.oas.models.OpenAPI;
 import org.openapitools.codegen.CodegenModel;
@@ -27,7 +29,7 @@ public class TwilioPhpGenerator extends PhpClientCodegen {
     private final List<CodegenModel> allModels = new ArrayList<>();
     private final Map<String, String> modelFormatMap = new HashMap<>();
     private IConventionMapper conventionMapper = new LanguageConventionResolver(PHP_CONVENTIONAL_MAP_PATH);
-
+    private IApiActionTemplate phpApiActionTemplate = new PhpApiActionTemplate(this);
     public TwilioPhpGenerator() {
         super();
         twilioCodegen = new TwilioCodegenAdapter(this, getName());
@@ -48,12 +50,11 @@ public class TwilioPhpGenerator extends PhpClientCodegen {
     public void processOpenAPI(final OpenAPI openAPI) {
         final String domain = twilioCodegen.getDomainFromOpenAPI(openAPI);
         twilioCodegen.setDomain(domain);
-        openAPI.getPaths().forEach(directoryStructureService.resourceTree::addResource);
-        directoryStructureService.resourceTree.getResources().forEach(resource -> resource.updateFamily(directoryStructureService.resourceTree));
+        directoryStructureService.configureResourceFamily(openAPI);
         directoryStructureService.configure(openAPI);
         if (!directoryStructureService.isVersionLess()) {
-            new PhpApiResourceBuilder(new PhpApiActionTemplate(this), new ArrayList<>(), new ArrayList<>())
-                    .setVersionTemplate(openAPI, domain, directoryStructureService);
+            new PhpDomainBuilder(phpApiActionTemplate)
+                    .setVersionTemplate(openAPI,directoryStructureService);
         }
     }
 
@@ -83,7 +84,7 @@ public class TwilioPhpGenerator extends PhpClientCodegen {
     }
 
     private ApiResources processCodegenOperations(List<CodegenOperation> opList) {
-        return new PhpApiResourceBuilder(new PhpApiActionTemplate(this), opList, this.allModels)
+        return new PhpApiResourceBuilder(phpApiActionTemplate, opList, this.allModels)
                 .updateApiPath()
                 .updateAdditionalProps(directoryStructureService)
                 .updateTemplate()
