+++ src/main/java/com/twilio/oai/template/IApiActionTemplate.java
@@ -3,4 +3,5 @@ package com.twilio.oai.template;
 public interface IApiActionTemplate {
     void clean();
     void add(String template);
+    void addSupportVersion();
 }
